+++ CONTEXT.md
@@ -322,16 +322,20 @@ The perfect-shape refactor is complete and merged. Its end-state:
   theirs.
 - Type-cycle growth (R9). R4 keeps the VALUE import graph acyclic, so every remaining cycle is
   created by type-only imports — free at runtime, invisible to R5/R6, and the largest single
-  obstacle to reading a subsystem in isolation: inside a strongly-connected component of 76 files,
+  obstacle to reading a subsystem in isolation: inside a strongly-connected component of 49 files,
   no file has a self-contained slice. `TYPE_CYCLE_BASELINE`, derived from the zone ceilings in
   `scripts/layering/daemon-modularity.ts`, ratchets it for **growth only**, deliberately unlike R6: reducing it
   is a real refactor rather than a file move, so a hard equality would turn every unrelated
   improvement into a baseline edit. A shrunk tree is reported in the success line instead of
-  failing. Hubs by in-component dependents: `runtime-contract.ts` (25),
-  `commands/runtime-types.ts` (21), `backend.ts` (15), `commands/runtime-common.ts` (12).
+  failing. Hubs by in-component dependents: `core/dispatch.ts` (8),
+  `command-catalog.ts` (7), `commands/interaction/runtime/resolution.ts` (6),
+  `core/command-descriptor/registry.ts` (6). The former type hubs (`runtime-contract.ts`,
+  `commands/runtime-types.ts`, `backend.ts`, `commands/runtime-common.ts`) left the cycle when
+  #1632 sank `backend.ts`'s two upward type imports — 27 files stranded out of the component at
+  once.
 - Daemon modularity ratchets (R10). The same tooling-only declaration pins R7's writer-owned
-  field/owner-claim counts, R9's 76 members by zone (`commands` 33, `daemon-server` 20, `core` 10,
-  `platforms` 7, root 5, `client` 1), and the external production importers of `daemon/types.ts`
+  field/owner-claim counts, R9's 49 members by zone (`commands` 14, `daemon-server` 19, `core` 10,
+  `platforms` 2, root 3, `client` 1), and the external production importers of `daemon/types.ts`
   (down to 2: the client normalizers and remote artifacts). R7 counts and external importers may
   only shrink; no zone may grow inside R9, and replay/Maestro/replay-test engine files remain outside
   it. This per-zone ratchet is intentionally stricter than R9's ordinary total-growth rule: even
