+++ src/utils/screenshot-result.ts
@@ -1,15 +1,8 @@
+import type { ScreenshotResultData } from '@agent-device/contracts/capture';
 import type { ScreenshotOverlayRef } from '@agent-device/kernel/snapshot';
 import { isRecord, parsePoint, parseRect } from './parsing.ts';
 
-export type ScreenshotResultData = {
-  path?: string;
-  width?: number;
-  height?: number;
-  logicalWidth?: number;
-  logicalHeight?: number;
-  pixelDensity?: number;
-  overlayRefs?: ScreenshotOverlayRef[];
-};
+export type { ScreenshotResultData } from '@agent-device/contracts/capture';
 
 export function pickScreenshotResultData(value: ScreenshotResultData): ScreenshotResultData {
   return {
